Update jsdocs for component
