+++ login-workflow/docs/components/set-password.md
@@ -19,8 +19,8 @@ import { SetPassword } from '@brightlayer-ui/react-native-auth-workflow';
 | Prop Name                     | Type                                                         | Description                                                                                                                     | Default |
 | ----------------------------- | ------------------------------------------------------------ | ------------------------------------------------------------------------------------------------------------------------------- | ------- |
 | onPasswordChange              | `(passwords: { password: string; confirm: string }) => void` | Called when the new password or confirm new password fields value changes.                                                      |         |
-| newPasswordLabel              | `string`                                                     | The label for the new password field.                                                                                           |         |
-| initialNewPasswordValue       | `string`                                                     | The initial value for the new password field.                                                                                   |         |
+| newPasswordLabel              | `string`                                                     | The label for the new password field.                                                                                           | 'Password' |
+| initialNewPasswordValue       | `string`                                                     | The initial value for the new password field.                                                                                   | 'Confirm' |
 | confirmPasswordLabel          | `string`                                                     | The label for the confirm password field.                                                                                       |         |
 | initialConfirmPasswordValue   | `string`                                                     | The initial value for the confirm password field.                                                                               |         |
 | passwordRequirements          | `PasswordRequirement[]`                                      | An array of password complexity requirements . See [PasswordRequirements](#PasswordRequirements) for more details.              |         |
