+++ login-workflow/src/components/Dialog/index.tsx
@@ -1,1 +1,2 @@
 export * from './BasicDialog';
+export * from './types';
