+++ login-workflow/src/components/Error/ErrorMessageBox.tsx
@@ -1,57 +1,14 @@
 import React from 'react';
-import { StyleProp, View, ViewStyle, StyleSheet, TextStyle } from 'react-native';
+import { View, ViewStyle, StyleSheet, TextStyle } from 'react-native';
 import { Text } from 'react-native-paper';
 import { ExtendedTheme, useExtendedTheme } from '@brightlayer-ui/react-native-themes';
 import Icon from 'react-native-vector-icons/MaterialIcons';
-
-export type ErrorMessageBoxProps = {
-    /**
-     * The text to show in the title
-     */
-    title: string;
-    /**
-     * The text to show in the message
-     */
-    errorMessage: string;
-
-    /**
-     * The background color of the message box
-     */
-    backgroundColor?: string;
-
-    /**
-     * Boolean whether the message box can be dismissed
-     * @default true
-     */
-    dismissible?: boolean;
-
-    /**
-     * The font color of the text inside the message box
-     */
-    fontColor?: string;
-
-    /**
-     * The function to call when the close button is clicked
-     * @returns void
-     */
-    onClose?: () => void;
-
-    /**
-     * Styles passed to the underlying root component
-     */
-    style?: StyleProp<ViewStyle>;
-};
+import { ErrorMessageBoxProps } from './types';
 
 /**
  * Component that renders a basic message box with an error message and a configurable dismiss button.
  *
- * @param text to show as the title
- * @param errorMessage text to show in the message
- * @param backgroundColor the background color of the message box
- * @param dismissible whether the message box can be dismissed
- * @param fontColor the font color of the text inside the message box
- * @param onClose function to call when the close button is clicked
- * @param style styles passed to the underlying root component
+ * @param {ErrorMessageBoxProps} props - Props of Error Message Box
  *
  * @category Component
  */
