+++ login-workflow/src/components/Error/index.tsx
@@ -1,2 +1,3 @@
 export * from './ErrorMessageBox';
 export * from './ErrorManager';
+export * from './types';
