+++ login-workflow/src/components/SetPassword/types.ts
@@ -7,36 +7,42 @@ import { TextInputProps } from 'react-native-paper';
  */
 export type PasswordRequirement = {
     /**
-     * The description of the password requirement
+     * @param {string} description - The description of the password requirement
+     *
      */
     description: string;
 
     /**
-     * The regex to validate the password
+     * @param {RegExp} regex - The regex to validate the password
+     *
      */
     regex: RegExp;
 };
 
 export type PasswordRequirementsProps = ViewProps & {
     /**
-     * The string to conduct the complexity checks against
+     * @param {string} passwordText - The string to conduct the complexity checks against
+     *
      */
     passwordText: string;
 
     /**
-     * Optional requirements to set password
+     * @param {PasswordRequirement[]} [passwordRequirements] - Optional requirements to set password
+     *
      */
     passwordRequirements?: PasswordRequirement[];
 };
 
 export type PasswordRequirementsCheckProps = {
     /**
-     * True if the line item should have a blue check (false for gray)
+     * @param {boolean} isChecked - True if the line item should have a blue check (false for gray)
+     *
      */
     isChecked: boolean;
 
     /**
-     * The text to display beside the check icon
+     * @param {string} label - The text to display beside the check icon
+     *
      */
     label: string;
 };
@@ -51,39 +57,44 @@ export type SetPasswordProps = {
     onPasswordChange?: (passwords: { password: string; confirm: string }) => void;
 
     /**
-     * The label for the new password field
-     * @default Password
+     *
+     * @param {string} [newPasswordLabel='Password'] - The label for the new password field
      */
     newPasswordLabel?: string;
 
     /**
-     * The initial value for the new password field
+     *
+     * @param {string} [initialNewPasswordValue] - The initial value for the new password field
      */
     initialNewPasswordValue?: string;
 
     /**
-     * The label for the confirm password field
-     * @default Confirm
+     *
+     * @param {string} [confirmPasswordLabel='Confirm'] - The label for the confirm password field
      */
     confirmPasswordLabel?: string;
 
     /**
-     * The initial value for the confirm password
+     *
+     * @param {string} [initialConfirmPasswordValue] - The initial value for the confirm password
      */
     initialConfirmPasswordValue?: string;
 
     /**
-     * A list of password complexity requirements
+     *
+     * @param {PasswordRequirement[]} [passwordRequirements] - A list of password complexity requirements
      */
     passwordRequirements?: PasswordRequirement[];
 
     /**
-     * The ref to forward to the password input
+     *
+     * @param {MutableRefObject<any>} [passwordRef] - The ref to forward to the password input
      */
     passwordRef?: MutableRefObject<any>;
 
     /**
-     * The ref to forward to the confirm password input
+     *
+     * @param {MutableRefObject<any>} [confirmRef] - The ref to forward to the confirm password input
      */
     confirmRef?: MutableRefObject<any>;
 
@@ -93,19 +104,19 @@ export type SetPasswordProps = {
     passwordNotMatchError?: string;
 
     /**
-     * The function to call when the form is submitted
+     * @param {() => void} [onSubmit] - The function to call when the form is submitted
      * @returns void
      */
     onSubmit?: () => void;
 
     /**
-     * The props to pass to the password field.
+     * @param {TextInputProps} [passwordTextFieldProps] - The props to pass to the password field.
      * See React Native Paper [TextInputProps API](https://callstack.github.io/react-native-paper/docs/components/TextInput/) for more details.
      */
     passwordTextFieldProps?: TextInputProps;
 
     /**
-     * The props to pass to the confirm password field.
+     * @param {TextInputProps} [confirmPasswordTextFieldProps] - The props to pass to the confirm password field.
      * See React Native Paper [TextInputProps API](https://callstack.github.io/react-native-paper/docs/components/TextInput/) for more details.
      */
     confirmPasswordTextFieldProps?: TextInputProps;
