+++ login-workflow/src/components/WorkflowCard/WorkflowCard.types.ts
@@ -4,11 +4,14 @@ import { CardTitleProps, TextProps } from 'react-native-paper';
 
 export type WorkflowCardBaseProps = ImageBackgroundProps & {
     /**
-     * If true, a blocking progress spinner will be displayed over the card
+     *
+     * @param {boolean} [loading=false] - If true, a blocking progress spinner will be displayed over the card
+     *
      */
     loading?: boolean;
     /**
-     * A custom background to render behind the card
+     *
+     * @param {ImageSourcePropType} [backgroundImage] - A custom background to render behind the card
      */
     backgroundImage?: ImageSourcePropType;
 };
@@ -17,89 +20,95 @@ export type WorkflowCardHeaderProps = CardTitleProps;
 
 export type WorkflowCardInstructionProps = Omit<TextProps<'bodyLarge'>, 'children' | 'theme' | 'variant'> & {
     /**
-     * The text to display as instructions
+     *
+     * @param {string | JSX.Element} [instructions] - The text to display as instructions
      */
     instructions?: string | JSX.Element;
     /**
-     * Whether or not to show a divider below the instructions
-     * @default true
+     *
+     * @param {boolean} [divider=true] - Whether or not to show a divider below the instructions
      */
     divider?: boolean;
 };
 
 export type WorkflowCardActionsProps = ViewProps & {
     /**
-     * True to display a divider above workflow card action buttons
-     * @default true
+     *
+     * @param {boolean} [divider=true] - True to display a divider above workflow card action buttons
+     *
      */
     divider?: boolean;
 
     /**
-     * Boolean or function that indicates whether the next button should be enabled
+     *
+     * @param {boolean | (() => boolean)} [canGoNext] - Boolean or function that indicates whether the next button should be enabled
      */
     canGoNext?: boolean | (() => boolean);
 
     /**
-     * Boolean or function that indicates whether the previous button should be enabled
+     *
+     * @param {boolean | (() => boolean)} [canGoPrevious] - Boolean or function that indicates whether the previous button should be enabled
      */
     canGoPrevious?: boolean | (() => boolean);
 
     /**
-     * Boolean that indicates whether the previous button should be displayed
+     *
+     * @param {boolean} [showPrevious=true] - Boolean that indicates whether the previous button should be displayed
      */
     showPrevious?: boolean;
 
     /**
-     * Boolean that indicates whether the next button should be displayed
+     *
+     * @param {boolean} [showNext=true] - Boolean that indicates whether the next button should be displayed
      */
     showNext?: boolean;
 
     /**
-     * The label to display for the previous button
+     *
+     * @param {string} [previousLabel] - The label to display for the previous button
      */
     previousLabel?: string;
 
     /**
-     * The label to display for the next button
+     *
+     * @param {string} [nextLabel] - The label to display for the next button
      */
     nextLabel?: string;
 
     /**
      * Function called when the previous button is clicked
-     * @param {Object} data - data collected from the card / screen
+     * @param {Object} [data] - data collected from the card / screen
      * @returns void
      */
     onPrevious?: (data?: { [key: string]: any }) => void;
 
     /**
      * Function that is called when the next button is clicked
-     * @param {Object} data - data collected from the card / screen
+     * @param {Object} [data] - data collected from the card / screen
      * @returns void
      */
     onNext?: (data?: { [key: string]: any }) => void;
 
     /**
-     * The current step in the registration workflow
-     * @default 0
+     *
+     * @param {number} [currentStep=0] - The current step in the registration workflow
      */
     currentStep?: number;
 
     /**
-     * The total number of steps in the registration workflow
-     * @default 5
+     *
+     * @param {number} [totalSteps=5] - The total number of steps in the registration workflow
      */
     totalSteps?: number;
 
     /**
-     * Boolean that indicates whether a button should be full width
+     *
+     * @param {boolean} [fullWidthButton=false] - Boolean that indicates whether a button should be full width
      */
     fullWidthButton?: boolean;
-    /** Which type of step indicator to use:
-     * - dots: circles
-     * - progress: progress bar
-     * - text: text
+    /**
      *
-     * @default: 'dots'
+     * @param {DotStepperVariant} [stepperVariant='dots'] - Which type of step indicator to use: dots: circles, progress: progress bar and text: text
      */
     stepperVariant?: DotStepperVariant;
 };
